requirements.txt
pymongo
discord
sympy
ossapi
pytube
pillow